chore: Storage Example
