+++ examples/wallet/src/components/object/info/index.tsx
@@ -1,4 +1,4 @@
-import { client } from '@/client';
+import { client, selectSp } from '@/client';
 import { useState } from 'react';
 
 export const ObjectInfo = () => {
@@ -25,6 +25,26 @@ export const ObjectInfo = () => {
         >
           get obj info
         </button>
+
+        <br />
+
+        <div>
+          get object by bucket name
+          <br />
+          <button
+            onClick={async () => {
+              const spInfo = await selectSp();
+
+              const res = await client.object.listObjects({
+                bucketName,
+                endpoint: spInfo.endpoint,
+              });
+              console.log('res', res);
+            }}
+          >
+            list objects
+          </button>
+        </div>
       </div>
     </div>
   );
